Commented out UserServiceTest, work still needed, implemented ProjectServiceTest #43
